Clean up code, update README commands, and fix issues in toy notebook
