+++ tsExperiments/utils/__init__.py
@@ -4,3 +4,4 @@ from utils.pylogger import RankedLogger
 from utils.rich_utils import enforce_tags, print_config_tree
 from utils.utils import extras, get_metric_value, task_wrapper
 from utils.utils import split_train_val
+from utils.utils import compute_metric_forecast
